+++ packages/atomic/cypress/integration/query-error-selectors.ts
@@ -1,5 +1,0 @@
-import {AriaLiveSelectors} from './aria-live-selectors';
-
-export const QueryErrorSelectors = {
-  liveRegion: () => AriaLiveSelectors.region('query-error'),
-};