+++ packages/quantic/cypress/integration/document-suggestion/document-suggestion-actions.ts
@@ -12,6 +12,18 @@ function documentSuggestionActions(selector: DocumentSuggestionSelector) {
         .logAction(
           `When clicking on the document suggestion at the following index: ${idx}`
         ),
+    openQuickview: (idx: number) =>
+      selector
+        .quickviewButton(idx)
+        .click()
+        .logAction(
+          `When opening the quickview of the document suggestion at the following index: ${idx}`
+        ),
+    closeQuickview: () =>
+      selector
+        .quickviewCloseButton()
+        .click()
+        .logAction('When closing the quickview of a document suggestion.'),
   };
 }