+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.46.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.45.2...@coveo/headless@1.46.0) (2022-02-11)
+
+
+### Features
+
+* **headless:** add originContext to analytics' state & configuration ([#1721](https://github.com/coveo/ui-kit/issues/1721)) ([e1974f8](https://github.com/coveo/ui-kit/commit/e1974f815c621604cdd0a8a2fbde87306a09c6a0))
+
+
+
+
+
+## [1.45.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.45.1...@coveo/headless@1.45.2) (2022-02-09)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+## [1.45.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.45.0...@coveo/headless@1.45.1) (2022-02-07)
+
+
+### Bug Fixes
+
+* **headless:** cleanup Jenkinsfile, add test for missing parsed_doc.json file ([#1708](https://github.com/coveo/ui-kit/issues/1708)) ([cd38df9](https://github.com/coveo/ui-kit/commit/cd38df9b1c24159b9a82aef08af45bed6623e072))
+* **headless, quantic:** Prevent search analytics event when using QV in case assist ([#1701](https://github.com/coveo/ui-kit/issues/1701)) ([4c3d63a](https://github.com/coveo/ui-kit/commit/4c3d63aeabed28fe9ca808eb8e1a640f2347143a))
+
+
+
+
+
+# [1.45.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.44.0...@coveo/headless@1.45.0) (2022-02-02)
+
+
+### Features
+
+* **headless:** allow toggleSelectFacetValue to add new values ([#1704](https://github.com/coveo/ui-kit/issues/1704)) ([403507e](https://github.com/coveo/ui-kit/commit/403507e38755c1164522e13f27fef17dd1fbc3f1))
+
+
+
+
+
 # [1.44.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.43.2...@coveo/headless@1.44.0) (2022-01-24)